django_server/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
decorator==5.2.1
Django==5.2.4
grpcio==1.73.1
grpcio-tools==1.73.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
service-identity==24.2.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
threadpoolctl==3.6.0
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webrtcvad==2.0.10
zope.interface==7.2
